Move routes and urls to common
