+++ frontends/mit-open/src/common/routes.tsx
@@ -1,17 +1,16 @@
 import React from "react"
 import { RouteObject, Outlet } from "react-router"
-import HomePage from "./pages/HomePage/HomePage"
+import HomePage from "pages/HomePage/HomePage"
 import RestrictedRoute from "components/RestrictedRoute/RestrictedRoute"
-import LearningPathListingPage from "./pages/LearningPathListingPage/LearningPathListingPage"
-import LearningPathDetailsPage from "./pages/LearningPathDetailsPage/LearningPathDetailsPage"
-import ArticleDetailsPage from "./pages/ArticleDetailsPage/ArticleDetailsPage"
-import { ArticleCreatePage, ArticleEditPage } from "./pages/ArticleUpsertPages"
-import ErrorPage from "./pages/ErrorPage/ErrorPage"
-import * as urls from "./pages/urls"
-import * as deprecatedUrls from "./pages/InfinitePages/common/infinite-pages-urls"
-import EditFieldPage from "./pages/InfinitePages/pages/EditFieldPage"
-import FieldPage from "./pages/InfinitePages/pages/FieldPage"
-
+import LearningPathListingPage from "pages/LearningPathListingPage/LearningPathListingPage"
+import LearningPathDetailsPage from "pages/LearningPathDetailsPage/LearningPathDetailsPage"
+import ArticleDetailsPage from "pages/ArticleDetailsPage/ArticleDetailsPage"
+import { ArticleCreatePage, ArticleEditPage } from "pages/ArticleUpsertPages"
+import ErrorPage from "pages/ErrorPage/ErrorPage"
+import * as urls from "./urls"
+import * as deprecatedUrls from "pages/InfinitePages/common/infinite-pages-urls"
+import EditFieldPage from "pages/InfinitePages/pages/EditFieldPage"
+import FieldPage from "pages/InfinitePages/pages/FieldPage"
 import Header from "components/Header/Header"
 import { Permissions } from "common/permissions"
