+++ frontends/mit-open/src/pages/ArticleDetailsPage/ArticleDetailsPage.tsx
@@ -6,7 +6,7 @@ import { useArticleDetail } from "api/hooks/articles"
 import { useParams } from "react-router"
 import { ButtonLink } from "ol-design"
 import { Grid } from "ol-design"
-import { articlesEditView } from "../urls"
+import { articlesEditView } from "common/urls"
 import { CkeditorDisplay } from "ol-ckeditor"
 
 type RouteParams = {
