+++ frontends/mit-open/src/pages/ArticleUpsertPages/ArticleEditPage.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from "react"
 import { useNavigate, useParams } from "react-router"
 import { useArticleDetail } from "api/hooks/articles"
 import ArticleUpsertForm from "page-components/ArticleUpsertForm/ArticleUpsertForm"
-import { articlesView } from "../urls"
+import { articlesView } from "common/urls"
 import ArticleUpsertPage from "./ArticleUpsertPage"
 
 type RouteParams = {
