+++ frontends/mit-open/src/services/api/learning-resources/resourceLists.test.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { renderHook, waitFor } from "@testing-library/react"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { faker } from "@faker-js/faker/locale/en"
-import { setMockResponse, act } from "../../test-utils"
+import { setMockResponse, act } from "../../../test-utils"
 import { invalidateResourceQueries } from "./util"
 import { keys, urls } from "./urls"
 import { LearningResourceType as LR, LearningResource } from "ol-search-ui"
