+++ frontends/mit-open/src/services/api/learning-resources/search.test.tsx
@@ -7,7 +7,7 @@ import { act } from "@testing-library/react"
 import { renderHook, waitFor } from "@testing-library/react"
 import * as factories from "ol-search-ui/src/test-utils/factories"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
-import { setMockResponse } from "../../test-utils/mockAxios"
+import { setMockResponse } from "../../../test-utils/mockAxios"
 import { urls } from "./urls"
 import axios from "services/axios"
 import { useInfiniteSearch } from "./search"
