+++ frontends/mit-open/src/services/api/learning-resources/util.test.tsx
@@ -2,7 +2,7 @@ import React from "react"
 import { renderHook, waitFor } from "@testing-library/react"
 import { QueryClient, QueryClientProvider } from "@tanstack/react-query"
 import { invalidateResourceQueries, useInfiniteLimitOffsetQuery } from "./util"
-import { setMockResponse, act } from "../../test-utils"
+import { setMockResponse, act } from "../../../test-utils"
 import axios from "services/axios"
 import {
   makeLearningResource,
