+++ frontends/mit-open/src/test-utils/index.tsx
@@ -2,11 +2,11 @@ import React from "react"
 import { createMemoryRouter, useRouteError } from "react-router"
 import type { RouteObject } from "react-router"
 
-import { AppProviders } from "../AppProviders"
-import appRoutes from "../routes"
+import App from "../App"
+import appRoutes from "common/routes"
 import { render } from "@testing-library/react"
 import { setMockResponse } from "./mockAxios"
-import { createQueryClient } from "../libs/react-query"
+import { createQueryClient } from "services/react-query/react-query"
 import type { User } from "../types/settings"
 import { makeUserSettings } from "./factories"
 
@@ -54,9 +54,7 @@ const renderRoutesWithProviders = (
     { initialEntries: [url] },
   )
   const queryClient = createQueryClient()
-  const view = render(
-    <AppProviders queryClient={queryClient} router={router}></AppProviders>,
-  )
+  const view = render(<App queryClient={queryClient} router={router}></App>)
 
   const location = {
     get current() {
